+++ src/components/molecules/BtnOptions.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script>
-import Btn from '../atoms/Btn'
+import Btn from '../ui/Btn'
 
 export default {
   components: {
